Fix PR comments.
