+++ src/common/constants.ts
@@ -4,6 +4,7 @@ export const INFO_SERVER_URI: string = 'https://info1.edgesecure.co:8444/v1'
 export const BYTES_TO_KB = '1000'
 
 // Fees
+export const FEES_PATH = 'fees.json'
 export const FEES_DISKLET_PATH = 'fees'
 export const MAX_FEE = 999999999.0
 export const LOW_FEE = 2
